Enforce only jsonb in check signature
Previously the signature expected from a check was of
the form (job_id, config). Also, there was no validation
that the check function given had the correct signature.
This commit removes the job_id as it is not required and
also checks that the check function has the correct signature
when it is registered with add_job, preventing an error being
thrown at job runtime.